src/requirements.txt
absl-py==0.13.0
addict==2.4.0
alembic==1.6.5
antlr4-python3-runtime==4.8
anyio==3.3.0
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
black==20.8b
bleach==4.0.0
cachetools==4.2.2
calmsize==0.1.3
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cliff==3.8.0
cmaes==0.8.2
cmd2==2.1.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
efficientnet-pytorch==0.6.3
entrypoints==0.3
flake8==3.8
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.34.0
google-auth-oauthlib==0.4.5
graphviz==0.17
greenlet==1.1.1
grpcio==1.39.0
idna==3.2
importlib-metadata==3.10.1
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jstyleson==0.0.2
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-server==1.10.2
jupyterlab==3.1.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
line-profiler==3.3.0
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multiprocess==0.70.12.2
munch==2.5.0
mypy
mysqlclient==2.0.3
natsort==7.1.1
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.2
ninja==1.10.2
nncf==2.0.0
notebook==6.4.2
numpy==1.19.5
oauthlib==3.1.1
omegaconf==2.1.0
optuna==2.9.1
packaging==21.0
pandas==1.3.1
pandocfilters==1.4.3
parso==0.8.2
pathos==0.2.8
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
Pillow-SIMD==7.0.0.post3
pox==0.3.0
ppft==1.6.6.4
pretrainedmodels==0.7.4
prettytable==2.1.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pylibjpeg-libjpeg==1.2.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
pysen
python-dateutil==2.8.2
python-editor==1.0.4
pytorch-memlab==0.2.3
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rich==10.7.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
segmentation-models-pytorch==0.2.0
Send2Trash==1.8.0
six==1.16.0
smmap==4.0.0
sniffio==1.2.0
SQLAlchemy==1.4.22
stevedore==3.3.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.1
testpath==0.5.0
texttable==1.6.4
threadpoolctl==2.2.0
timm==0.4.12
toml==0.10.2
torch==1.9.0
torch-tb-profiler==0.2.1
torchaudio==0.9.0
torchinfo==1.5.3
torchvision==0.10.0
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.0
Werkzeug==2.0.1
widgetsnbextension==3.5.1
zipp==3.5.0
